Remove qstash call from links API, rearrange links API
