+++ apps/web/lib/api/links/process-link.ts
@@ -1,0 +1,260 @@
+import {
+  isBlacklistedDomain,
+  isBlacklistedKey,
+  isReservedKey,
+  isReservedUsername,
+} from "@/lib/edge-config";
+import prisma from "@/lib/prisma";
+import { isStored, storage } from "@/lib/storage";
+import { formatRedisLink, redis } from "@/lib/upstash";
+import {
+  getLinksCountQuerySchema,
+  getLinksQuerySchema,
+} from "@/lib/zod/schemas/links";
+import {
+  APP_DOMAIN_WITH_NGROK,
+  DEFAULT_REDIRECTS,
+  DUB_DOMAINS,
+  SHORT_DOMAIN,
+  getDomainWithoutWWW,
+  getParamsFromURL,
+  getUrlFromString,
+  isDubDomain,
+  isValidUrl,
+  linkConstructor,
+  truncate,
+  validKeyRegex,
+} from "@dub/utils";
+import { Prisma } from "@prisma/client";
+import { checkIfKeyExists, getRandomKey } from "@/lib/planetscale";
+import { recordLink } from "@/lib/tinybird";
+import {
+  LinkProps,
+  LinkWithTagIdsProps,
+  RedisLinkProps,
+  WorkspaceProps,
+} from "@/lib/types";
+import z from "@/lib/zod";
+import { qstash } from "@/lib/cron";
+import { combineTagIds, keyChecks, processKey } from "./utils";
+
+export async function processLink({
+  payload,
+  workspace,
+  userId,
+  bulk = false,
+  skipKeyChecks = false, // only skip when key doesn't change (e.g. when editing a link)
+}: {
+  payload: LinkWithTagIdsProps;
+  workspace?: WorkspaceProps;
+  userId?: string;
+  bulk?: boolean;
+  skipKeyChecks?: boolean;
+}) {
+  let {
+    domain,
+    key,
+    url,
+    image,
+    proxy,
+    password,
+    rewrite,
+    expiresAt,
+    ios,
+    android,
+    geo,
+  } = payload;
+
+  const tagIds = combineTagIds(payload);
+
+  // url checks
+  if (!url) {
+    return {
+      link: payload,
+      error: "Missing destination url.",
+      code: "bad_request",
+    };
+  }
+  const processedUrl = getUrlFromString(url);
+  if (!isValidUrl(processedUrl)) {
+    return {
+      link: payload,
+      error: "Invalid destination url.",
+      code: "unprocessable_entity",
+    };
+  }
+
+  // free plan restrictions
+  if (!workspace || workspace.plan === "free") {
+    if (proxy || password || rewrite || expiresAt || ios || android || geo) {
+      return {
+        link: payload,
+        error:
+          "You can only use custom social media cards, password-protection, link cloaking, link expiration, device and geo targeting on a Pro plan and above. Upgrade to Pro to use these features.",
+        code: "forbidden",
+      };
+    }
+  }
+
+  // if domain is not defined, set it to the workspace's primary domain
+  if (!domain) {
+    domain = workspace?.domains?.find((d) => d.primary)?.slug || SHORT_DOMAIN;
+  }
+
+  // checks for default short domain
+  if (domain === SHORT_DOMAIN) {
+    const keyBlacklisted = await isBlacklistedKey(key);
+    if (keyBlacklisted) {
+      return {
+        link: payload,
+        error: "Invalid key.",
+        code: "unprocessable_entity",
+      };
+    }
+    const domainBlacklisted = await isBlacklistedDomain(url);
+    if (domainBlacklisted) {
+      return {
+        link: payload,
+        error: "Invalid url.",
+        code: "unprocessable_entity",
+      };
+    }
+
+    // checks for other Dub-owned domains (chatg.pt, spti.fi, etc.)
+  } else if (isDubDomain(domain)) {
+    // coerce type with ! cause we already checked if it exists
+    const { allowedHostnames } = DUB_DOMAINS.find((d) => d.slug === domain)!;
+    const urlDomain = getDomainWithoutWWW(url) || "";
+    if (!allowedHostnames.includes(urlDomain)) {
+      return {
+        link: payload,
+        error: `Invalid url. You can only use ${domain} short links for URLs starting with ${allowedHostnames
+          .map((d) => `\`${d}\``)
+          .join(", ")}.`,
+        code: "unprocessable_entity",
+      };
+    }
+
+    // else, check if the domain belongs to the workspace
+  } else if (!workspace?.domains?.find((d) => d.slug === domain)) {
+    return {
+      link: payload,
+      error: "Domain does not belong to workspace.",
+      code: "forbidden",
+    };
+  }
+
+  if (!key) {
+    key = await getRandomKey(domain, payload["prefix"]);
+  } else if (!skipKeyChecks) {
+    const processedKey = processKey(key);
+    if (processedKey === null) {
+      return {
+        link: payload,
+        error: "Invalid key.",
+        code: "unprocessable_entity",
+      };
+    }
+    key = processedKey;
+
+    const response = await keyChecks({ domain, key, workspace });
+    if (response.error) {
+      return {
+        link: payload,
+        ...response,
+      };
+    }
+  }
+
+  if (bulk) {
+    if (image) {
+      return {
+        link: payload,
+        error: "You cannot set custom social cards with bulk link creation.",
+        code: "unprocessable_entity",
+      };
+    }
+    if (rewrite) {
+      return {
+        link: payload,
+        error: "You cannot use link cloaking with bulk link creation.",
+        code: "unprocessable_entity",
+      };
+    }
+
+    // only perform tag validity checks if:
+    // - not bulk creation (we do that check separately in the route itself)
+    // - tagIds are present
+  } else if (tagIds.length > 0) {
+    const tags = await prisma.tag.findMany({
+      select: {
+        id: true,
+      },
+      where: { projectId: workspace?.id, id: { in: tagIds } },
+    });
+
+    if (tags.length !== tagIds.length) {
+      return {
+        link: payload,
+        error:
+          "Invalid tagIds detected: " +
+          tagIds
+            .filter(
+              (tagId) => tags.find(({ id }) => tagId === id) === undefined,
+            )
+            .join(", "),
+        code: "unprocessable_entity",
+      };
+    }
+  }
+
+  // custom social media image checks (see if R2 is configured)
+  if (proxy && !process.env.STORAGE_SECRET_ACCESS_KEY) {
+    return {
+      link: payload,
+      error: "Missing storage access key.",
+      code: "bad_request",
+    };
+  }
+
+  // expire date checks
+  if (expiresAt) {
+    const date = new Date(expiresAt);
+    if (isNaN(date.getTime())) {
+      return {
+        link: payload,
+        error: "Invalid expiry date. Expiry date must be in ISO-8601 format.",
+        code: "unprocessable_entity",
+      };
+    }
+    // check if expiresAt is in the future
+    if (new Date(expiresAt) < new Date()) {
+      return {
+        link: payload,
+        error: "Expiry date must be in the future.",
+        code: "unprocessable_entity",
+      };
+    }
+  }
+
+  // remove polyfill attributes from payload
+  delete payload["shortLink"];
+  delete payload["qrCode"];
+  delete payload["prefix"];
+
+  return {
+    link: {
+      ...payload,
+      domain,
+      key,
+      url: processedUrl,
+      // make sure projectId is set to the current workspace
+      projectId: workspace?.id || null,
+      // if userId is passed, set it (we don't change the userId if it's already set, e.g. when editing a link)
+      ...(userId && {
+        userId,
+      }),
+    },
+    error: null,
+  };
+}
