+++ apps/web/vercel.json
@@ -7,6 +7,10 @@
     {
       "path": "/api/cron/usage",
       "schedule": "0 14 * * *"
+    },
+    {
+      "path": "/api/links/usage",
+      "schedule": "*/5 * * * *"
     }
   ],
   "functions": {
